backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
pymysql
pydantic<2.0.0
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
